Battery Voltage Testing

+++ Controller/Controller.ino
@@ -15,7 +15,7 @@ String mesgDictionary[] = {
 //create an RF24 object
 RF24 radio(9, 8);  // CE, CSN
 const byte address[5] = {'R','x','A','A','A'};
-int ackData[2] = {-1, -1};
+float ackData[2] = {-1, -1};
 //bool newAck = false;
 
 // Max size of this struct is 32 bytes - NRF24L01 buffer limit
